CardFloatingView 분리
